+++ consensus/core/src/network.rs
@@ -1,4 +1,4 @@
-use borsh::{BorshDeserialize, BorshSchema, BorshSerialize};
+use borsh::{BorshDeserialize, BorshSerialize};
 use kaspa_addresses::Prefix;
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use std::fmt::{Debug, Display, Formatter};
@@ -15,9 +15,7 @@ pub enum NetworkTypeError {
 }
 
 /// @category Consensus
-#[derive(
-    Clone, Copy, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, BorshSchema, PartialEq, Eq, Hash, Ord, PartialOrd,
-)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq, Hash, Ord, PartialOrd)]
 #[serde(rename_all = "lowercase")]
 #[wasm_bindgen]
 pub enum NetworkType {
@@ -182,7 +180,7 @@ impl From<NetworkIdError> for JsValue {
 ///
 /// @category Consensus
 ///
-#[derive(Clone, Copy, Debug, BorshSerialize, BorshDeserialize, BorshSchema, PartialEq, Eq, Hash, Ord, PartialOrd, CastFromJs)]
+#[derive(Clone, Copy, Debug, BorshSerialize, BorshDeserialize, PartialEq, Eq, Hash, Ord, PartialOrd, CastFromJs)]
 #[wasm_bindgen(inspectable)]
 pub struct NetworkId {
     #[wasm_bindgen(js_name = "type")]