+++ wallet/core/src/utxo/processor.rs
@@ -6,6 +6,7 @@
 //!
 
 use crate::imports::*;
+// use futures::pin_mut;
 use kaspa_notify::{
     listener::ListenerId,
     scope::{Scope, UtxosChangedScope, VirtualDaaScoreChangedScope},
@@ -28,11 +29,12 @@ use crate::utxo::{
     Maturity, OutgoingTransaction, PendingUtxoEntryReference, SyncMonitor, UtxoContext, UtxoEntryId, UtxoEntryReference,
 };
 use crate::wallet::WalletBusMessage;
-use async_std::sync::{Mutex as AsyncMutex, MutexGuard as AsyncMutexGuard};
 use kaspa_rpc_core::{
     notify::connection::{ChannelConnection, ChannelType},
     Notification,
 };
+// use workflow_core::task;
+// use kaspa_metrics_core::{Metrics,Metric};
 
 pub struct Inner {
     /// Coinbase UTXOs in stasis
@@ -56,7 +58,10 @@ pub struct Inner {
     sync_proc: SyncMonitor,
     multiplexer: Multiplexer<Box<Events>>,
     wallet_bus: Option<Channel<WalletBusMessage>>,
-    notification_lock: AsyncMutex<()>,
+    notification_guard: AsyncMutex<()>,
+    connect_disconnect_guard: AsyncMutex<()>,
+    metrics: Arc<Metrics>,
+    metrics_kinds: Mutex<Vec<MetricsUpdateKind>>,
 }
 
 impl Inner {
@@ -82,7 +87,10 @@ impl Inner {
             sync_proc: SyncMonitor::new(rpc.clone(), &multiplexer),
             multiplexer,
             wallet_bus,
-            notification_lock: AsyncMutex::new(()),
+            notification_guard: Default::default(),
+            connect_disconnect_guard: Default::default(),
+            metrics: Arc::new(Metrics::default()),
+            metrics_kinds: Mutex::new(vec![]),
         }
     }
 }
@@ -121,8 +129,14 @@ impl UtxoProcessor {
 
     pub async fn bind_rpc(&self, rpc: Option<Rpc>) -> Result<()> {
         *self.inner.rpc.lock().unwrap() = rpc.clone();
+        let rpc_api = rpc.as_ref().map(|rpc| rpc.rpc_api().clone());
+        self.metrics().bind_rpc(rpc_api);
         self.sync_proc().bind_rpc(rpc).await?;
         Ok(())
+    }
+
+    pub fn metrics(&self) -> &Arc<Metrics> {
+        &self.inner.metrics
     }
 
     pub fn wallet_bus(&self) -> &Option<Channel<WalletBusMessage>> {
@@ -138,7 +152,7 @@ impl UtxoProcessor {
     }
 
     pub async fn notification_lock(&self) -> AsyncMutexGuard<()> {
-        self.inner.notification_lock.lock().await
+        self.inner.notification_guard.lock().await
     }
 
     pub fn sync_proc(&self) -> &SyncMonitor {
@@ -450,22 +464,43 @@ impl UtxoProcessor {
         self.notify(Events::UtxoProcStart).await?;
         self.sync_proc().track(is_synced).await?;
 
+        let this = self.clone();
+        self.inner.metrics.register_sink(Arc::new(Box::new(move |snapshot: MetricsSnapshot| {
+            if let Err(err) = this.deliver_metrics_snapshot(Box::new(snapshot)) {
+                println!("Error ingesting metrics snapshot: {}", err);
+            }
+            None
+        })));
+
         Ok(())
     }
 
     pub async fn handle_connect(&self) -> Result<()> {
-        if let Err(err) = self.handle_connect_impl().await {
-            log_error!("UtxoProcessor: error while connecting to node: {err}");
-            self.notify(Events::UtxoProcError { message: err.to_string() }).await?;
-            if let Some(client) = self.rpc_client() {
-                client.disconnect().await?;
+        let _ = self.inner.connect_disconnect_guard.lock().await;
+
+        match self.handle_connect_impl().await {
+            Err(err) => {
+                log_error!("UtxoProcessor: error while connecting to node: {err}");
+                self.notify(Events::UtxoProcError { message: err.to_string() }).await?;
+                if let Some(client) = self.rpc_client() {
+                    // try force disconnect the client if we have failed
+                    // to negotiate the connection to the node.
+                    client.disconnect().await?;
+                }
+                Err(err)
             }
+            Ok(_) => Ok(()),
         }
-        Ok(())
     }
 
     pub async fn handle_disconnect(&self) -> Result<()> {
+        let _ = self.inner.connect_disconnect_guard.lock().await;
+
         self.inner.is_connected.store(false, Ordering::SeqCst);
+        // self.stop_metrics();
+
+        self.inner.metrics.unregister_sink();
+
         self.unregister_notification_listener().await?;
         self.notify(Events::UtxoProcStop).await?;
         self.cleanup().await?;
@@ -486,9 +521,7 @@ impl UtxoProcessor {
             .rpc_api()
             .register_new_listener(ChannelConnection::new(self.inner.notification_channel.sender.clone(), ChannelType::Persistent));
         *self.inner.listener_id.lock().unwrap() = Some(listener_id);
-
         self.rpc_api().start_notify(listener_id, Scope::VirtualDaaScoreChanged(VirtualDaaScoreChangedScope {})).await?;
-
         Ok(())
     }
 
@@ -525,6 +558,37 @@ impl UtxoProcessor {
         Ok(())
     }
 
+    fn deliver_metrics_snapshot(&self, snapshot: Box<MetricsSnapshot>) -> Result<()> {
+        let metrics_kinds = self.inner.metrics_kinds.lock().unwrap().clone();
+        for kind in metrics_kinds.into_iter() {
+            match kind {
+                MetricsUpdateKind::WalletMetrics => {
+                    let mempool_size = snapshot.get(&Metric::NetworkMempoolSize) as u64;
+                    let node_peers = snapshot.get(&Metric::NodeActivePeers) as u32;
+                    let network_tps = snapshot.get(&Metric::NetworkTransactionsPerSecond);
+                    let metrics = MetricsUpdate::WalletMetrics { mempool_size, node_peers, network_tps };
+                    self.try_notify(Events::Metrics { network_id: self.network_id()?, metrics })?;
+                }
+            }
+        }
+
+        Ok(())
+    }
+
+    pub async fn start_metrics(&self) -> Result<()> {
+        self.inner.metrics.start_task().await?;
+        self.inner.metrics.bind_rpc(Some(self.rpc_api().clone()));
+
+        Ok(())
+    }
+
+    pub async fn stop_metrics(&self) -> Result<()> {
+        self.inner.metrics.stop_task().await?;
+        self.inner.metrics.bind_rpc(None);
+
+        Ok(())
+    }
+
     pub async fn start(&self) -> Result<()> {
         let this = self.clone();
         if this.inner.task_is_running.load(Ordering::SeqCst) {
@@ -554,11 +618,14 @@ impl UtxoProcessor {
                                 match msg {
                                     RpcState::Opened => {
                                         if !this.is_connected() {
-                                            this.handle_connect().await.unwrap_or_else(|err| log_error!("{err}"));
-                                            this.inner.multiplexer.try_broadcast(Box::new(Events::Connect {
-                                                network_id : this.network_id().expect("network id expected during connection"),
-                                                url : this.rpc_url()
-                                            })).unwrap_or_else(|err| log_error!("{err}"));
+                                            if let Err(err) = this.handle_connect().await {
+                                                log_error!("UtxoProcessor error: {err}");
+                                            } else {
+                                                this.inner.multiplexer.try_broadcast(Box::new(Events::Connect {
+                                                    network_id : this.network_id().expect("network id expected during connection"),
+                                                    url : this.rpc_url()
+                                                })).unwrap_or_else(|err| log_error!("{err}"));
+                                            }
                                         }
                                     },
                                     RpcState::Closed => {
@@ -605,6 +672,8 @@ impl UtxoProcessor {
                 }
             }
 
+            log_info!("utxo processing task existing");
+
             // handle power down on rpc channel that remains connected
             if this.is_connected() {
                 this.handle_disconnect().await.unwrap_or_else(|err| log_error!("{err}"));
@@ -622,6 +691,10 @@ impl UtxoProcessor {
             self.inner.task_ctl.signal(()).await.expect("UtxoProcessor::stop_task() `signal` error");
         }
         Ok(())
+    }
+
+    pub fn enable_metrics_kinds(&self, metrics_kinds: &[MetricsUpdateKind]) {
+        *self.inner.metrics_kinds.lock().unwrap() = metrics_kinds.to_vec();
     }
 }