+++ wallet/core/src/wasm/utxo/processor.rs
@@ -56,6 +56,27 @@ impl Inner {
     }
 }
 
+cfg_if! {
+    if #[cfg(any(feature = "wasm32-core", feature = "wasm32-sdk"))] {
+        #[wasm_bindgen(typescript_custom_section)]
+        const TS_NOTIFY: &'static str = r#"
+        interface UtxoProcessor {
+            /**
+            * @param {UtxoProcessorNotificationCallback} callback
+            */
+            addEventListener(callback:UtxoProcessorNotificationCallback): void;
+            /**
+            * @param {UtxoProcessorEventType} event
+            * @param {UtxoProcessorNotificationCallback} [callback]
+            */
+            addEventListener<M extends keyof UtxoProcessorEventMap>(
+                event: M,
+                callback: (eventData: UtxoProcessorEventMap[M]) => void
+            )
+        }"#;
+    }
+}
+
 ///
 /// UtxoProcessor class is the main coordinator that manages UTXO processing
 /// between multiple UtxoContext instances. It acts as a bridge between the
@@ -196,9 +217,23 @@ impl UtxoProcessor {
                             let callbacks = inner.callbacks(event_type);
                             if let Some(handlers) = callbacks {
                                 for handler in handlers.into_iter() {
-                                    let value = to_value(&notification).unwrap();
-                                    if let Err(err) = handler.call(&value) {
-                                        log_error!("Error while executing RPC notification callback: {:?}", err);
+                                    match notification.as_ref() {
+                                        Events::Pending { record } |
+                                        Events::Reorg { record } |
+                                        Events::Stasis { record } |
+                                        Events::Maturity { record } |
+                                        Events::Discovery { record } => {
+                                            let value = JsValue::from(record.clone());
+                                            if let Err(err) = handler.call(&value) {
+                                                log_error!("Error while executing notification callback: {:?}", err);
+                                            }
+                                        }
+                                        _ => {
+                                            let value = to_value(&notification).unwrap();
+                                            if let Err(err) = handler.call(&value) {
+                                                log_error!("Error while executing notification callback: {:?}", err);
+                                            }
+                                        }
                                     }
                                 }
                             }
@@ -226,7 +261,7 @@ impl UtxoProcessor {
 
 #[wasm_bindgen]
 impl UtxoProcessor {
-    #[wasm_bindgen(js_name = "addEventListener")]
+    #[wasm_bindgen(js_name = "addEventListener", skip_typescript)]
     pub fn add_event_listener(
         &self,
         event: UtxoProcessorNotificationTypeOrCallback,