+++ wasm/examples/nodejs/typescript/src/utxo-context-listener.ts
@@ -1,5 +1,5 @@
 // Run with: node demo.js
-import {w3cwebsocket} from "websocket";
+import { w3cwebsocket } from "websocket";
 (globalThis.WebSocket as any) = w3cwebsocket;
 
 import {
@@ -13,9 +13,14 @@ import {
     createTransactions,
     initConsolePanicHook,
     IUtxoProcessorEvent,
+    IPendingEvent,
+    IDiscoveryEvent,
+    IMaturityEvent,
+    ITransactionRecord,
+    UtxoProcessorEventType,
 } from "../../../../nodejs/kaspa";
 
-import {parseArgs} from "./utils";
+import { parseArgs } from "./utils";
 
 
 initConsolePanicHook();
@@ -38,7 +43,7 @@ let { encoding, networkId, destinationAddress } = parseArgs();
 
     // 1) Initialize RPC
     const rpc = new RpcClient({
-        resolver : new Resolver(),
+        resolver: new Resolver(),
         encoding,
         networkId
     });
@@ -53,8 +58,35 @@ let { encoding, networkId, destinationAddress } = parseArgs();
     let context = new UtxoContext({ processor });
 
     // 4) Register a listener with the UtxoProcessor::events
-    processor.addEventListener((event:IUtxoProcessorEvent) => {
-        console.log("event:", event);
+    processor.addEventListener(({event, data}) => {
+        // handle the event
+        // since the event is a union type, you can switch on the 
+        // event type and cast the data to the appropriate type
+        switch (event) {
+            //string enums are not working for @Matoo
+            //case UtxoProcessorEventType.Discovery: {
+            case "discovery": {
+                let record = data.record;
+                console.log("Discovery event record:", record);
+            } break;
+            //case UtxoProcessorEventType.Pending: {
+            case "pending":{
+                let record = data.record;
+                console.log("Pending event record:", record);
+            } break;
+            //case UtxoProcessorEventType.Maturity: {
+            case "maturity": {
+                let record = data.record;
+                console.log("Maturity event record:", record);
+            } break;
+            default: {
+                console.log("Other event:", event, "data:", data);
+            }
+        }
+    });
+
+    processor.addEventListener("discovery", (data) => {
+        console.log("Discovery event record:", data.record);
     });
 
     console.log(processor);